Fix nasty bug if there's an error in the callback
If you universalify.fromPromise() a function, and pass a callback in
that contains an error (as is common with mocha tests), universalify
would call the callback again with the error when the callback threw.

This commit attempts to fix this as best we can; now errors will cause
an unhandled promise rejection (which is OK, because normally a
callback error causes an unhandled rejection).