Optionally allow false as a "no error" value
Even though not all falsey values should be considered as non-errors
which was fixed in #13, there are built-in callbacks which use `false`
as the no-error value. (eg `fs.realpath`)

The current state breaks `fs-extra` which uses universalify to wrap
`fs.realpath` ending up with successful calls raising an exception.

In this PR I am adding an option to consider "false" error values as
"no-error".